refactor(agent): split profile.ts into identity, capabilities, and runtime modules
